backend/pyproject.toml
[project]
name = "dvw-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ole Bause",email = "ole@bause.me"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.0,<6.0)",
    "djangorestframework (>=3.15,<4.0)",
    "psycopg2-binary (>=2.9,<3.0)",
    "django-cors-headers (>=4.3,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.7"
pytest = "^8.4.1"
pytest-django = "^4.11.1"

